PMS Assignment changes

+++ PMS/Models/Project.cs
@@ -12,6 +12,6 @@ namespace PMS.Models
         public string Description { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int Budget { get; set; }
+        public double Budget { get; set; }
     }
 }
